Move VoyagerConfig definition into its own model file.
