+++ src/models/config.ts
@@ -1,0 +1,7 @@
+export interface VoyagerConfig {
+  showDataSourceSelector?: boolean;
+};
+
+export const DEFAULT_VOYAGER_CONFIG = {
+  showDataSourceSelector: false
+};
